requirements.txt
openai==0.26.4
pandas
plotly
scipy
scikit-learn==1.2.1
pyarrow
gradio
python-dotenv==1.0.1
altair